borked image_model user uploading... working on it

+++ application/models/image_model.php
@@ -21,19 +21,23 @@ class Image_model extends CI_Model
 
 	function make_profile_images($upload_file, $upload_width, $upload_height, $user_id)
 	{
+		// Load library, helper
 		$this->load->helper('file');
 	    $this->load->library('image_lib');		
-				
+		
+		// Make Folder / Delete Existing		
 		make_folder(config_item('users_images_folder').$user_id);		
 		delete_files(config_item('users_images_folder').$user_id."/");
 	    
-	    $raw_path 		= config_item('uploads_folder').$upload_file;
-	    $thumb_path 	= config_item('users_images_folder').$user_id."/".$upload_file;
+	    // Paths
+	    $raw_path 			= config_item('uploads_folder').$upload_file;
+	    $thumb_path 		= config_item('users_images_folder').$user_id."/".$upload_file;
+//	    $square_path 		= config_item('users_images_folder').$user_id."/square_".$upload_file;
 	    
 	    $original_width		= 0;
 	    $original_height	= 0;
 	    
-	    // Raw width larger than allowed
+	    // Raw Width Larger Than Allowed
 		if ($upload_width >= config_item('users_images_full_width'))
 		{
 			$original_width = config_item('users_images_full_width');
@@ -43,7 +47,7 @@ class Image_model extends CI_Model
 			$original_width = $upload_width;
 		}
 
-	    // Raw height larger than allowed
+	    // Raw Height Larger Than Allowed
 		if ($upload_height >= config_item('users_images_full_height'))
 		{
 			$original_height = config_item('users_images_full_height');
@@ -85,11 +89,11 @@ class Image_model extends CI_Model
 	        $y_axis = round($diff / 2);
 	    }
 	    
-	    // Makes largest size possible square image	 
+	    // Largest Possible Square	 
 		$crop_config['image_library']	= 'gd2';
 	    $crop_config['source_image'] 	= $raw_path;
 	    $crop_config['maintain_ratio']	= FALSE;
-	    $crop_config['new_image'] 		= config_item('users_images_folder').$user_id."/".$upload_file; 
+	    $crop_config['new_image'] 		= config_item('users_images_folder').$user_id."/square_".$upload_file; 
 	    $crop_config['x_axis']		 	= $x_axis;
 	    $crop_config['y_axis'] 			= $y_axis;
 	    $crop_config['width'] 			= $cropsize;
@@ -108,8 +112,8 @@ class Image_model extends CI_Model
   	    
 
 		// Full image crop resize
-		if (config_item('users_images_sizes_medium') == 'yes')
-		{  	    
+		if (config_item('users_images_sizes_full') == 'yes')
+		{  				
 		    $resize_config['image_library'] 	= 'gd2';
 		    $resize_config['source_image'] 		= $raw_path;
 		    $resize_config['maintain_ratio'] 	= TRUE;
@@ -132,8 +136,10 @@ class Image_model extends CI_Model
 
 
 		// Large image crop resize
-		if (config_item('users_images_sizes_medium') == 'yes')
+		if (config_item('users_images_sizes_large') == 'yes')
 		{
+			//if (config_item('users_images_large_width') == config_item('users_images_large_height')) $thumb_path = $square_path;
+		
 		    $thumb_config['image_library'] 		= 'gd2';
 		    $thumb_config['source_image'] 		= $thumb_path;
 		    $thumb_config['maintain_ratio'] 	= TRUE;
@@ -156,12 +162,14 @@ class Image_model extends CI_Model
 		// Medium image crop resize
 		if (config_item('users_images_sizes_medium') == 'yes')
 		{
+			//if (config_item('users_images_medium_width') == config_item('users_images_medium_height')) $thumb_path = $square_path;
+		
 		    $thumb2_config['image_library'] 	= 'gd2';
 		    $thumb2_config['source_image'] 		= $thumb_path;
 		    $thumb2_config['maintain_ratio'] 	= TRUE;
 		    $thumb2_config['new_image']			= config_item('users_images_folder').$user_id."/"."medium_".$upload_file;
 		    $thumb2_config['width'] 			= config_item('users_images_medium_width');
-		    $thumb2_config['height'] 			= config_item('users_images_medium_height');
+		    $thumb2_config['height'] 			= config_item('users_images_medium_height');		    
 		    
 		    $this->image_lib->initialize($thumb2_config);
 		    
@@ -178,12 +186,14 @@ class Image_model extends CI_Model
 		// Small image crop resize
 		if (config_item('users_images_sizes_small') == 'yes')
 		{
+			//if (config_item('users_images_small_width') == config_item('users_images_small_height')) $thumb_path = $square_path;		
+		
 		    $thumb3_config['image_library'] 	= 'gd2';
 		    $thumb3_config['source_image'] 		= $thumb_path;
 		    $thumb3_config['maintain_ratio'] 	= TRUE;
 		    $thumb3_config['new_image']			= config_item('users_images_folder').$user_id."/"."small_".$upload_file;
 		    $thumb3_config['width'] 			= config_item('users_images_small_width');
-		    $thumb3_config['height'] 			= config_item('users_images_small_height');
+		    $thumb3_config['height'] 			= config_item('users_images_small_height');		    
 		    
 		    $this->image_lib->initialize($thumb3_config);
 		    
@@ -195,7 +205,14 @@ class Image_model extends CI_Model
 		    }
 		}
 	    
-	    unlink($thumb_path);
+	    
+		// Medium image crop resize
+		if (config_item('users_images_sizes_original') == 'no')
+		{	    
+	    	unlink($thumb_path);
+	    }
+
+	   // unlink($square_path);
 	    
 	    return true;
